+++ numlab/compiler/__init__.py
@@ -1,6 +1,11 @@
 from numlab.compiler.generic_ast import AST
 from numlab.compiler.grammar import (Grammar, NonTerminal, Production, Symbol,
                                      Terminal)
-from numlab.compiler.parser import Parser
+from numlab.compiler.parsers.parser import Parser
+from numlab.compiler.parsers.lritem import LRItem
+from numlab.compiler.parsers.lr1_parser import LR1Parser, LR1Table
+
+from numlab.compiler.parser_manager import ParserManager
+
 from numlab.compiler.terminal_set import TerminalSet
 from numlab.compiler.tokenizer import Token, Tokenizer
